+++ Changelog.md
@@ -1,3 +1,14 @@
+# v1.0.0-beta.2
+
+## Additions
+
+-   Added a progress bar.
+-   Added color to the console output.
+
+## Fixes
+
+-   Fixed a bug where the script would crash if the `panoramas` folder contained any files or empty folders.
+
 # v1.0.0-beta.1
 
 -   Initial release